changd DPOAEAnlyzer to handle spectrum with type double

+++ trunk/src/org/audiopulse/analysis/DPOAEAnalyzer.java
@@ -48,7 +48,7 @@ public class DPOAEAnalyzer {
 
 	private final static String TAG="DPOAEAnalyzer";
 	public static final String TestType="DPOAE";
-	private int[] XFFT;
+	private double[] XFFT;
 	private final static double spectralToleranceHz=50;
 	private double Fs;
 	private double F1;
@@ -59,7 +59,7 @@ public class DPOAEAnalyzer {
 	//Estimated levels above the threshold below will be logged as errors
 	private static final double dBErrorWarningThreshold=70;
 
-	public DPOAEAnalyzer(int [] XFFT, double Fs, double F2, double F1, double Fres,
+	public DPOAEAnalyzer(double [] XFFT, double Fs, double F2, double F1, double Fres,
 			String protocol, String fileName){
 		this.Fs=Fs;
 		this.XFFT=XFFT;
@@ -94,7 +94,7 @@ public class DPOAEAnalyzer {
 		return dResults;
 	}
 
-	private double[][] getSpectrum(int[] xFFT2, double fs) {
+	private double[][] getSpectrum(double[] xFFT2, double fs) {
 		// Reformat data to two arrays where the first is the frequency index
 		double[][] PFFT=new double[2][xFFT2.length];
 		double step= (double) fs/(2.0*(xFFT2.length-1));
